+++ test/data/conversion-length.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-lines */
 import { ConversionTest } from "./conversion-test";
-import * as Units from "../../src/units";
+import { Units } from "../../src/all-units";
 
 export const tests: ReadonlyArray<ConversionTest> = [
   {
